refactor: move most of Checkout variables and helpers methods to utils and libs folder
