--- src/components/checkout/helpers/variables.ts
@@ -1,5 +1,0 @@
-export const textMatchRegexp = /\D/g
-
-export const zipCodeRegexp = /^(\d{5})(\d{1,3})$/g
-
-export const creditCardRegexp = /(\d{4})/g
