+++ src/libs/zod/errors.ts
@@ -6,7 +6,3 @@ export function lengthErrorMessage(fieldName: string, type: MinOrMax, value: num
     }
     return `${fieldName} must have at least ${value} character(s)`
 }
-
-export function requiredErrorMessage(fieldName: string): string {
-    return `${fieldName} is required`
-}
