+++ src/utils/variables.ts
@@ -2,6 +2,10 @@ import HeadphoneThumb from "/public/imgs/headphones-thumbnail.png"
 import SpeakerThumb from "/public/imgs/speakers-thumbnail.png"
 import EarphoneThumb from "/public/imgs/earphones-thumbnail.png"
 
+export const textMatchRegexp = /\D/g
+export const zipCodeRegexp = /^(\d{5})(\d{1,3})$/g
+export const creditCardRegexp = /(\d{4})/g
+
 export const categories = [
     {
         name: "headphones",
